# Conflicts:
#	backend/src/main/java/edu/netcracker/backend/controller/ServiceController.java
#	frontend/src/main/web/src/app/modules/api/index.ts
#	frontend/src/main/web/src/app/pages/carrier/carrier.module.ts
#	frontend/src/main/web/src/app/pages/carrier/services/archive/archive.component.ts
#	frontend/src/main/web/src/app/pages/carrier/services/clarification/clarification.component.ts
#	frontend/src/main/web/src/app/pages/carrier/services/service-crud/service-crud.component.ts
#	frontend/src/main/web/src/app/pages/carrier/services/service-table/service-table.component.ts
#	frontend/src/main/web/src/app/pages/carrier/services/shared/service/service.service.ts

+++ backend/src/main/java/edu/netcracker/backend/service/AdminService.java
@@ -5,10 +5,10 @@ import java.time.LocalDateTime;
 import java.util.Map;
 
 public interface AdminService {
-    Map<Integer, Integer> getCostsPerPeriodPerCarrier(Number id, LocalDateTime from, LocalDateTime to);
-    Map<Integer, Integer> getCostsPerPeriod(LocalDateTime from, LocalDateTime to);
+    Map<Float, Long> getCostsPerPeriodPerCarrier(Number id, LocalDateTime from, LocalDateTime to);
+    Map<Float, Long> getCostsPerPeriod(LocalDateTime from, LocalDateTime to);
 
-    Map<LocalDateTime, Integer> getUsersIncreasingPerPeriod(LocalDateTime from, LocalDateTime to);
-    Map<LocalDateTime, Integer> getCarriersIncreasingPerPeriod(LocalDateTime from, LocalDateTime to);
-    Map<LocalDateTime, Integer> getLocationsIncreasingPerPeriod(LocalDateTime from, LocalDateTime to);
+    Map<LocalDateTime, Long> getUsersIncreasingPerPeriod(LocalDateTime from, LocalDateTime to);
+    Map<LocalDateTime, Long> getCarriersIncreasingPerPeriod(LocalDateTime from, LocalDateTime to);
+    Map<LocalDateTime, Long> getLocationsIncreasingPerPeriod(LocalDateTime from, LocalDateTime to);
 }
